% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{adc}
\alias{adc}
\title{Anomaly Detection and Correction}
\usage{
adc(data, eu = "in", pva = "BY", alpha = 0.01)
}
\arguments{
\item{data}{A \code{data.frame} with colnames "id", "x", "y".}

\item{eu}{Character string specifying the residual test.  Must be "ex" or
"in".}

\item{pva}{Character string specifying p-value adjustment.  See pea_adjust.}

\item{alpha}{Numeric value of nominal significance level.}
}
\value{
Original data with an additional columns: "ana" (logical on whether the
observation is an anomaly), "rep" (replacement value, if necessary), and "apv" (adjusted p-value).
}
\description{
Anomaly Detection and Correction
}
\examples{
# To do.
}
