% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{create.data}
\alias{create.data}
\title{Create anomalous data}
\usage{
create.data(fun, id, zeta = 1, sigma = 0.1, n = 50)
}
\arguments{
\item{fun}{A character specifying functional form.  Options include "Beta",
"Cos", "Exp" and "Quad".}

\item{id}{A unique subject identifier.}

\item{zeta}{Numeric; magnitude of anomaly.}

\item{sigma}{Numeric; noise.}

\item{n}{Numeric; sample size.}
}
\value{
A data.frame with columns, id, x, f, z and y.
}
\description{
Creates data for one subject.  User can specify functional form, anomaly
size, error and sample size.
}
\examples{
set.seed(808)
dat1 <- rbind(create.data("Beta", 1),
              create.data("Cos", 2, 0),
              create.data("Exp", 3, 2),
              create.data("Quad", 4))
plot(y ~ x, data = dat1, col = id)
}
