% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{internal}
\alias{internal}
\title{Eubank's internally standardized residual test}
\usage{
internal(mod)
}
\arguments{
\item{mod}{A \code{gamObject} from \strong{mgcv}.}
}
\value{
a list with p-values, diagonal of hat matrix and their sum (aka.
degrees of freedom), and the test statistic values.
}
\description{
Figure out how to make LaTeX-like formulas.
}
\examples{
# to do
}
