% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\name{magic2gam}
\alias{magic2gam}
\title{Update \code{gamObject} after \code{magic}}
\usage{
magic2gam(gam_F, magic, w)
}
\arguments{
\item{gam_F}{A \code{gamObject} with \code{fit = FALSE}.}

\item{magic}{Results from \code{magic}.}

\item{w}{The (decomposed) information matrix.}
}
\value{
A \code{gamObject}.
}
\description{
Update \code{gamObject} after \code{magic}
}
\examples{
# to do
}
