% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\name{meth.external}
\alias{meth.external}
\title{Implement the external test}
\usage{
meth.external(data, method, sig = 0.01)
}
\arguments{
\item{data}{A \code{data.frame} with colnames "id", "x", "y".}

\item{method}{Character.  See \code{pea_adjust}.}

\item{sig}{Numeric.  Nominal significance level.}
}
\value{
\code{data} with an additional column "ana", logical on whether the
observation is an anomaly.
}
\description{
The external test is wrapped in a \code{for} loop.  Fits a smoothing spline under
AR-1.  (Need to generalize to other corr structures.)
}
\examples{
# to do
}
