% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preplot.R
\name{preplot}
\alias{preplot}
\title{Prepare results for plotting}
\usage{
preplot(res)
}
\arguments{
\item{res}{A \code{data.frame} output from \code{\link[=adc]{adc()}}.}
}
\value{
A \code{data.frame} that is longer and wider so that the suggested values
can be plotted with a different color and symbol.
}
\description{
Reorganizes the output from \code{\link[=adc]{adc()}} so it can be plotted.  It sorts the
subjects based their smallest p-values.
}
\examples{
data("GermanHyperinflation")
dat <- data.frame(id = "a",
                  x = log(GermanHyperinflation$pi),
                  y = GermanHyperinflation$logM)
res  <- adc(data = dat, eu = "in", pva = "none", alpha = 0.05)
res2 <- preplot(res = res)
plot(y ~ x, data = res2, col = col, pch = pch)
}
