% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adc.R
\name{adc}
\alias{adc}
\title{Anomaly detection and correction}
\usage{
adc(data, eu = "in", pva = "BY", alpha = 0.01)
}
\arguments{
\item{data}{A \code{data.frame} with column names "id" (subject identifier), "x"
(predictor) and "y" (response).}

\item{eu}{Character string specifying the residual test.  Must be "ex" or
"in".}

\item{pva}{Character string specifying p-value adjustment.  See \code{\link[=pea_adjust]{pea_adjust()}}.}

\item{alpha}{Numeric value of nominal significance level.}
}
\value{
Original data with additional columns: "ana" (logical on whether the
observation is an anomaly), "rep" (replacement value, if necessary), and
"apv" (adjusted p-value).
}
\description{
\code{adc()} detects and corrects anomalies in functional data.  It sports two
types of residuals (internal and external) for smoothing splines, and a
variety of multiple comparison adjustments.
}
\examples{
data("GermanHyperinflation")
dat <- data.frame(id = 1,
                  x = log(GermanHyperinflation$pi),
                  y = GermanHyperinflation$logM)
res <- adc(data = dat, eu = "in", pva = "none", alpha = 0.05)
}
\references{
Kuwaye and Cho, 2024
}
\seealso{
\code{\link[=pea_adjust]{pea_adjust()}}.
}
