% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{external}
\alias{external}
\title{Externally standardized residual test}
\usage{
external(mod, x)
}
\arguments{
\item{mod}{A \code{gamObject} from \strong{mgcv}.}

\item{x}{The independent variable, usually time.}
}
\value{
A list with p-values, diagonal of hat matrix and their sum (aka.
degrees of freedom), and the test statistic values.
}
\description{
The formulas in the paper are incorrect.  The correct formulas are below.
}
\examples{
# To do.
}
