% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{internal}
\alias{internal}
\title{Internally standardized residual test}
\usage{
internal(mod, x)
}
\arguments{
\item{mod}{A \code{gamObject} from \strong{mgcv}.}

\item{x}{The independent variable, usually time.}
}
\value{
A list with p-values, diagonal of hat matrix and their sum (aka.
degrees of freedom), and the test statistic values.
}
\description{
Internally standardized residual test
}
\examples{
# To do.
}
